Example mission showcasing moving respawn tents
For actual use remove the kill trigger, it's just there to quickly force you into the respawn screen. In order to add more respawns, you can simple duplicate the trigger and tent. Then edit the tent variable name and adjust that in the code to reference the new tent and to remove the old respawn. Variables should all be pretty obvious but if not you can dm me